вынесены утилитарные функции
